briefly/collaboration_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic-settings==2.1.0
